requirements.txt
lxml==4.6.4
beautifulsoup4==4.10.0
selenium==4.1.0
requests==2.26.0
Pillow==8.4.0
pytesseract==0.3.8
mitmproxy==7.0.4
Faker==2.0.2
numpy==1.21.5